Move startGraphQLServer code from cli to recipes
